feat: Refactor board coordinates to use `u8` for improved memory efficiency and update related logic.

+++ tests/castling_general.rs
@@ -4,11 +4,11 @@ use hyperchess::domain::models::{Piece, PieceType, Player};
 use hyperchess::domain::rules::Rules;
 
 fn coord_2d(x: usize, y: usize) -> Coordinate {
-    Coordinate::new(vec![x, y])
+    Coordinate::new(vec![x as u8, y as u8])
 }
 
 fn coord_3d(x: usize, y: usize, z: usize) -> Coordinate {
-    Coordinate::new(vec![x, y, z])
+    Coordinate::new(vec![x as u8, y as u8, z as u8])
 }
 
 #[test]
